feat: Implementación de Canonical JSON
